+++ Scroller/EditableWrapper.js
@@ -210,7 +210,7 @@ const EditableWrapper = (props) => {
 	}, [customCss.noBefore, customCss.noAfter]);
 
 	const startEditing = useCallback((item) => {
-		if (item?.dataset?.index && (!item.hasAttribute('disabled') || item.className.includes('hidden'))) {
+		if (item?.dataset?.index && (!item.hasAttribute('disabled') || item.hasAttribute('data-is-hiding'))) {
 			item.classList.add(componentCss.selected, customCss.selected);
 			mutableRef.current.selectedItem = item;
 			mutableRef.current.focusedItem?.classList.remove(customCss.focused);
@@ -272,7 +272,10 @@ const EditableWrapper = (props) => {
 	}, [customCss.focused, findItemNode]);
 
 	const handleClickCapture = useCallback((ev) => {
-		if (ev.target?.parentNode?.parentNode.getAttribute('role') === 'button') {
+		const isButtonBg = ev.target.getAttribute('role') === 'button';
+		const isButtonClient = ev.target?.parentNode.getAttribute('role') === 'button';
+		const isButtonIcon = ev.target?.parentNode?.parentNode.getAttribute('role') === 'button';
+		if (isButtonBg || isButtonClient || isButtonIcon) {
 			return;
 		}
 		// Consume the event to prevent Item behavior
@@ -284,7 +287,10 @@ const EditableWrapper = (props) => {
 	}, []);
 
 	const handleMouseDown = useCallback((ev) => {
-		if (ev.target?.parentNode?.parentNode.getAttribute('role') === 'button') {
+		const isButtonBg = ev.target.getAttribute('role') === 'button';
+		const isButtonClient = ev.target?.parentNode.getAttribute('role') === 'button';
+		const isButtonIcon = ev.target?.parentNode?.parentNode.getAttribute('role') === 'button';
+		if (isButtonBg || isButtonClient || isButtonIcon) {
 			return;
 		}
 		if (mutableRef.current.selectedItem) {
@@ -390,7 +396,7 @@ const EditableWrapper = (props) => {
 		const {selectedItem} = mutableRef.current;
 		const {rtl} = scrollContainerHandle.current;
 
-		if (selectedItem && !selectedItem.className.includes('hidden')) {
+		if (selectedItem && !selectedItem.hasAttribute('data-is-hiding')) {
 			// Bail out when index is out of scope
 			if (toIndex < mutableRef.current.hideIndex && toIndex >= 0) {
 				const {fromIndex, itemWidth, moveDirection, prevToIndex, rearrangedItems} = mutableRef.current;
@@ -506,6 +512,7 @@ const EditableWrapper = (props) => {
 				item.style.order -= 1;
 			});
 			targetItem.style.order = orders.length;
+			targetItem.setAttribute('data-is-hiding', true);
 
 			finalizeEditing(orders);
 		}
@@ -525,6 +532,7 @@ const EditableWrapper = (props) => {
 			orders.splice(mutableRef.current.hideIndex, 0, targetItemOrder);
 
 			mutableRef.current.hideIndex += 1;
+			targetItem.removeAttribute('data-is-hiding');
 
 			finalizeEditing(orders);
 		}