+++ VideoPlayer/VideoPlayer.js
@@ -29,7 +29,6 @@ import Media from '@enact/ui/Media';
 import Slottable from '@enact/ui/Slottable';
 import Touchable from '@enact/ui/Touchable';
 import DurationFmt from 'ilib/lib/DurationFmt';
-import equals from 'ramda/src/equals';
 import PropTypes from 'prop-types';
 import {cloneElement, Component, createRef, isValidElement} from 'react';
 
@@ -245,6 +244,15 @@ const VideoPlayerBase = class extends Component {
 		 * @public
 		 */
 		feedbackHideDelay: PropTypes.number,
+
+		/**
+		 * Checks if current time and total time should include the hour.
+		 *
+		 * @type {Boolean}
+		 * @default false
+		 * @public
+		 */
+		includeTimeHour: PropTypes.bool,
 
 		/**
 		 * Components placed below the title.
@@ -778,6 +786,7 @@ const VideoPlayerBase = class extends Component {
 	static defaultProps = {
 		autoCloseTimeout: 5000,
 		feedbackHideDelay: 3000,
+		includeTimeHour: false,
 		initialJumpDelay: 400,
 		jumpBy: 30,
 		jumpDelay: 200,
@@ -878,13 +887,6 @@ const VideoPlayerBase = class extends Component {
 	}
 
 	componentDidUpdate (prevProps, prevState) {
-		if (this.titleRef && this.state.infoVisible &&
-			(!prevState.infoVisible || !equals(this.props.infoComponents, prevProps.infoComponents))
-		) {
-			const node = this.titleRef;
-			node.style.setProperty('--infoComponentsOffset', `${node.offsetHeight}px`);
-		}
-
 		if (
 			!this.state.mediaControlsVisible && prevState.mediaControlsVisible !== this.state.mediaControlsVisible ||
 			!this.state.mediaSliderVisible && prevState.mediaSliderVisible !== this.state.mediaSliderVisible
@@ -1990,6 +1992,7 @@ const VideoPlayerBase = class extends Component {
 			noSlider,
 			noSpinner,
 			selection,
+			includeTimeHour,
 			spotlightDisabled,
 			spotlightId,
 			style,
@@ -2122,7 +2125,7 @@ const VideoPlayerBase = class extends Component {
 										{infoComponents}
 									</MediaTitle>
 									{noSlider ?
-										<Times current={this.state.currentTime} total={this.state.duration} formatter={durFmt} /> :
+										<Times current={this.state.currentTime} total={this.state.duration} formatter={durFmt} includeHour={includeTimeHour} /> :
 										null
 									}
 								</div> :
@@ -2132,7 +2135,7 @@ const VideoPlayerBase = class extends Component {
 								null :
 								<div className={css.sliderContainer}>
 									{this.state.mediaSliderVisible ?
-										<Times noTotalTime current={this.state.currentTime} formatter={durFmt} /> :
+										<Times noTotalTime current={this.state.currentTime} formatter={durFmt} includeHour={includeTimeHour}  /> :
 										null
 									}
 									<MediaSlider
@@ -2162,7 +2165,7 @@ const VideoPlayerBase = class extends Component {
 										/>
 									</MediaSlider>
 									{this.state.mediaSliderVisible ?
-										<Times noCurrentTime total={this.state.duration} formatter={durFmt} /> :
+										<Times noCurrentTime total={this.state.duration} formatter={durFmt} includeHour={includeTimeHour} /> :
 										null
 									}
 								</div>