+++ samples/sampler/stories/default/Card.js
@@ -1,8 +1,8 @@
 import {Card, CardBase} from '@enact/limestone/Card';
 import {mergeComponentMetadata} from '@enact/storybook-utils';
+import {action} from '@enact/storybook-utils/addons/actions';
 import {boolean, object, select, text} from '@enact/storybook-utils/addons/controls';
 import {Card as UiCard} from '@enact/ui/Card';
-import ri from '@enact/ui/resolution';
 
 import {svgGenerator} from '../helper/svg';
 
@@ -29,12 +29,15 @@ export default {
 export const _Card = (args) => (
 	<Card
 		captionOverlay={args['captionOverlay']}
+		captionOverlayOnFocus={args['captionOverlayOnFocus']}
 		centered={args['centered']}
 		disabled={args['disabled']}
 		imageIconSrc={args['imageIconSrc']}
+		imageSize={args['imageSize']}
 		hasContainer={args['hasContainer']}
 		// eslint-disable-next-line no-undefined
 		label={args['label'] ? args['label'] : undefined}
+		onClick={action('onClick')}
 		orientation={args['orientation']}
 		primaryBadgeSrc={args['primaryBadgeSrc']}
 		roundedImage={args['roundedImage']}
@@ -43,20 +46,17 @@ export const _Card = (args) => (
 		secondaryLabel={args['secondaryLabel'] ? args['secondaryLabel'] : undefined}
 		selected={args['selected']}
 		src={args['src']}
-		style={{
-			position: 'absolute',
-			width: ri.scaleToRem(args['orientation'] === 'vertical' ? 768 : 1500),
-			height: ri.scaleToRem(args['orientation'] === 'vertical' ? 708 : 336)
-		}}
 	>
 		{args['children']}
 	</Card>
 );
 
 boolean('captionOverlay', _Card, Config);
+boolean('captionOverlayOnFocus', _Card, Config);
 boolean('centered', _Card, Config);
 boolean('disabled', _Card, Config);
 object('imageIconSrc', _Card, Config, generateImageSrc('0084ff'));
+object('imageSize', _Card, Config);
 boolean('hasContainer', _Card, Config);
 text('label', _Card, Config, 'Card label');
 select('orientation', _Card, prop.orientation, Config);