+++ samples/sampler/stories/default/TabLayout.js
@@ -7,7 +7,7 @@ import Scroller from '@enact/limestone/Scroller';
 import TabLayout, {TabLayoutBase, Tab} from '@enact/limestone/TabLayout';
 import {mergeComponentMetadata} from '@enact/storybook-utils';
 import {action} from '@enact/storybook-utils/addons/actions';
-import {range, select} from '@enact/storybook-utils/addons/controls';
+import {boolean, range, select} from '@enact/storybook-utils/addons/controls';
 import {scaleToRem} from '@enact/ui/resolution';
 
 import {svgGenerator} from '../helper/svg';
@@ -59,6 +59,7 @@ export const _TabLayout = (args) => {
 		<Panel>
 			<Header title="Limestone TabLayout" subtitle="Basic TabLayout" />
 			<TabLayout
+				blockCollapseOnPortrait={args['blockCollapseOnPortrait']}
 				onSelect={action('onSelect')}
 				onTabAnimationEnd={action('onTabAnimationEnd')}
 				orientation={args['orientation']}
@@ -95,10 +96,12 @@ export const _TabLayout = (args) => {
 	);
 };
 
+boolean('blockCollapseOnPortrait', _TabLayout, Config);
 select('tabs', _TabLayout, ['with icons', 'without icons'], Config, 'with icons');
 select('size', _TabLayout, ['small', 'large'], Config, 'large');
 select('orientation', _TabLayout, ['vertical', 'horizontal'], Config);
-range('tabSize', _TabLayout, Config, {min: 0, max: 960, step: 60}, 0);
+boolean('custom tabSize', _TabLayout, Config, false);
+range('tabSize', _TabLayout, Config, {min: 300, max: 960, step: 60}, 300, {if: {arg: 'custom tabSize'}});
 
 _TabLayout.storyName = 'TabLayout';
 _TabLayout.parameters = {